+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/metrics/RepositoryGauges.scala
@@ -30,7 +30,7 @@ class RepositoryGauges(metricRegistry: MetricRegistry,
       implicit val user: LoggedUser = NussknackerInternalUser
       val result = processRepository.fetchProcessesDetails[Unit](FetchProcessesDetailsQuery(isArchived = Some(false))).map { scenarios =>
         val all = scenarios.size
-        val deployed = scenarios.count(_.isDeployed)
+        val deployed = scenarios.count(_.state.exists(_.status.isDeployed))
         val fragments = scenarios.count(_.isSubprocess)
         Values(all, deployed, fragments)
       }
