+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/helpers/MockFetchingProcessRepository.scala
@@ -29,7 +29,7 @@ class MockFetchingProcessRepository(processes: List[BaseProcessDetails[_]])(impl
 
   override def fetchProcessesDetails[PS: ProcessShapeFetchStrategy](q: FetchProcessesDetailsQuery)(implicit loggedUser: LoggedUser, ec: ExecutionContext): Future[List[BaseProcessDetails[PS]]] =
     getUserProcesses[PS].map(_.filter(
-      p => check(q.isSubprocess, p.isSubprocess) && check(q.isArchived, p.isArchived) && check(q.isDeployed, p.isDeployed) && checkSeq(q.categories, p.processCategory) && checkSeq(q.processingTypes, p.processingType)
+      p => check(q.isSubprocess, p.isSubprocess) && check(q.isArchived, p.isArchived) && check(q.isDeployed, p.state.exists(_.status.isDeployed)) && checkSeq(q.categories, p.processCategory) && checkSeq(q.processingTypes, p.processingType)
     ))
 
   override def fetchLatestProcessDetailsForProcessId[PS: ProcessShapeFetchStrategy](id: ProcessId)(implicit loggedUser: LoggedUser, ec: ExecutionContext): Future[Option[BaseProcessDetails[PS]]] =
